+++ src/wetterdienst/provider/fmi/observation/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (C) 2018-2025, earthobservations developers.
+# Distributed under the MIT License. See LICENSE for more info.
+from wetterdienst.provider.fmi.observation.api import FmiObservationRequest
+from wetterdienst.provider.fmi.observation.metadata import FmiObservationMetadata
+
+__all__ = ["FmiObservationMetadata", "FmiObservationRequest"]